refactor(RacingGame at el): Change package structure
