+++ src/main/kotlin/racingcar/Main.kt
@@ -1,12 +1,12 @@
 package racingcar
 
-import racingcar.communication.ControlTower
-import racingcar.communication.Pit
-import racingcar.communication.input.Input
-import racingcar.communication.input.InputConsole
-import racingcar.communication.output.Output
-import racingcar.communication.output.OutputConsole
-import racingcar.domain.RacingGame
+import racingcar.util.ControlTower
+import racingcar.util.Pit
+import racingcar.view.input.Input
+import racingcar.view.input.InputConsole
+import racingcar.view.output.Output
+import racingcar.view.output.OutputConsole
+import racingcar.controller.RacingGame
 import racingcar.domain.cars.RacingCars
 import racingcar.domain.engine.Engine
