+++ src/main/kotlin/racingcar/util/Pit.kt
@@ -1,7 +1,7 @@
-package racingcar.communication
+package racingcar.util
 
-import racingcar.communication.input.Input
-import racingcar.communication.output.Output
+import racingcar.view.input.Input
+import racingcar.view.output.Output
 import racingcar.domain.cars.RacingCar
 
 object Pit {
